feat: 서버 관리 추가
