+++ Hansori/lib/server.dart
@@ -1,0 +1,9 @@
+import 'dart:io';
+
+class Server {
+  String name;
+  List<String> invitedFriends;
+  File? image;
+
+  Server({required this.name, required this.invitedFriends, this.image});
+}
